todo crud api

+++ src/todos/todo.controller.ts
@@ -2,28 +2,43 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   NotFoundException,
   Param,
   Post,
   Put,
+  Request,
+  UseGuards,
 } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
 import { ApiTags } from '@nestjs/swagger';
 import { plainToClass } from 'class-transformer';
 import { DeleteResult } from 'typeorm/index';
 import { EntityId } from 'typeorm/repository/EntityId';
-import { TodoDto } from './todo.dto';
+import {
+  TodoCreateReqDto,
+  TodoDto,
+  TodosDto,
+  TodoUpdateReqDto,
+} from './todo.dto';
 import { TodoService } from './todo.service';
 
 @ApiTags('todos')
 @Controller('todos')
+@UseGuards(AuthGuard('jwt'))
 export class TodoController {
-  // constructor(private moduleRef: ModuleRef) {}
   constructor(private readonly todoService: TodoService) {}
 
   @Get()
-  index(): Promise<TodoDto[]> {
-    return this.todoService.index();
+  async showAll(@Request() request): Promise<TodosDto> {
+    const { id } = request.user;
+    const todos = await this.todoService.findByUserId(id);
+    return plainToClass(
+      TodosDto,
+      { data: todos[0], totalCount: todos[1] },
+      { excludeExtraneousValues: true },
+    );
   }
 
   @Get('/:id')
@@ -39,23 +54,49 @@ export class TodoController {
   }
 
   @Post()
-  async create(@Body() userData: TodoDto): Promise<TodoDto> {
-    const createdUser = await this.todoService.save(userData);
+  async create(
+    @Request() request,
+    @Body() userData: TodoCreateReqDto,
+  ): Promise<TodoDto> {
+    const { id } = request.user;
+    const createdUser = await this.todoService.save({
+      ...userData,
+      userId: id,
+    });
     return plainToClass(TodoDto, createdUser, {
       excludeExtraneousValues: true,
     });
   }
 
   @Put('/:id')
-  update(
+  async update(
+    @Request() request,
     @Param('id') id: EntityId,
-    @Body() userData: TodoDto,
+    @Body() data: TodoUpdateReqDto,
   ): Promise<TodoDto> {
-    return this.todoService.update(id, userData);
+    const todo = await this.todoService.findById(id);
+    if (!todo) {
+      throw new NotFoundException();
+    }
+    if (todo.userId != request.user.id) {
+      throw new ForbiddenException();
+    }
+
+    return this.todoService.update(id, data);
   }
 
   @Delete('/:id')
-  destroy(@Param('id') id: EntityId): Promise<DeleteResult> {
+  async destroy(
+    @Request() request,
+    @Param('id') id: EntityId,
+  ): Promise<DeleteResult> {
+    const todo = await this.todoService.findById(id);
+    if (!todo) {
+      throw new NotFoundException();
+    }
+    if (todo.userId != request.user.id) {
+      throw new ForbiddenException();
+    }
     return this.todoService.delete(id);
   }
 }
